backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-dotenv==1.0.1
openai==1.30.1
sentence-transformers==2.7.0
chromadb==0.5.3
pypdf==4.2.0
python-docx==1.1.2
langchain-text-splitters==0.2.1
python-multipart==0.0.9
pydantic==2.7.1
pydantic-settings==2.3.0
pytest==8.2.2
httpx==0.27.0
pytest-asyncio==0.23.7
